ui: refactor; extract JobStatusCell and JobTable components
Moving towards a single-component-per-file design that's easier
to manage and modify

Release note: none
